requirements.txt
asgiref==3.9.0
asttokens==3.0.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-csp==4.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
executing==2.2.0
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
keyboard==0.13.5
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.4.2
numba==0.61.2
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-decouple==3.8
regex==2024.11.6
requests==2.32.3
setuptools==80.4.0
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
